OS X 10.4 SDK @ Xcode 2.2.1
Filename: MacOSX10.4u.sdk

http://adcdownload.apple.com/Developer_Tools/xcode_tools_2.2.1/xcode_2.2.1_8g1165_018213632.dmg

+++ MacOSXA.B.C.sdk/System/Library/Frameworks/vecLib.framework/Versions/A/Headers/cblas.h
@@ -694,7 +694,7 @@ void SetBLASParamErrorProc(BLASParamErrorProc ErrorProc);
     #endif
 
 #elif defined(__i386__)
-    typedef int __M128  __attribute__((vector_size (16)));
+    typedef float __M128  __attribute__((vector_size (16)));
 
     typedef __M128  				VectorFloat;
     typedef VectorFloat				ConstVectorFloat;
